Fix issues caused by merge with cocos2d 3.8. It's NOT FUNCTIONAL yet.

+++ cocos/base/CCDirector.cpp
@@ -1397,7 +1397,7 @@ void DisplayLinkDirector::stopAnimation()
     _invalid = true;
 }
 
-void DisplayLinkDirector::setAnimationInterval(double interval)
+void DisplayLinkDirector::setAnimationInterval(float interval)
 {
     _animationInterval = interval;
     if (! _invalid)
